Reverts all commits related to generating navmeshes in gamelogic
Fix #1854

The reason to no longer try to debug but instead do a bulk revert are
the following ones:

* every time I find a new bug in that code, I'm loosing several
  hours checking if it does not comes from my own code
* the PR litterally only adds bugs, no features nor improvements of any
  sort
* the PR is even used as a pretext to slow down integration of actual
  improvements in pathfinding
* the PR does not follows any coding style already used in gamelogic
* there is no guarantee the feature actually works as no server
  demonstrated yet AFAIK

Reverted commits:
* bbe18d3dce6082f685e7fcf81ca563bf4e24e3a9.
  "Fully revert lazy navmesh loading"
* fb8e18626e46b2c2a941b936a071b7b40e94c3a0.
  "Fix wrong return in navgen error handling"
* ffcd67ab87f7c3797ea57c57be9390482658e170.
  "partial revert of 44b29a6 to fix a pathfinding regression"
* 3d106269c95d46da5239c81333c83c80acec1f9d.
  "Load navmeshes on 'navedit enable'"
* d804c034f546e0c6c3fb469699dce3a3eb7c06c8.
  "Deduplicate navmesh file path building"
* e5c417dc0baa2a08174443463b8a9ea5b5376dcb.
  "Tweak navgen log levels"
* 0c1c8e6be38cfb66f4c8144c11cab0c8bb007d9f.
  "'navgen all' and 'navgen missing' commands"
* 668c65a886f60af38dde6397bc3c6197fd4a45f1.
  "Error handling for navmesh generation"
* 231d1618b37a1003bfd0b80ac6195e9ad61bcad7.
  "Preparation for better cache invalidation"
* a005e7e52c02c455fb6a34c2bd1f847518f5ed59.
  "Hard-code list of classes requiring navmesh"
* 44b29a6c45ce23c6afb0142371411fec761ce249.
  "Init navmeshes only when bot is added"
* b790448e539da2e6c1d735d83859c83b2ce63879.
  "Generate navmeshes during cgame loading screen"
* 89dbd5187bf7829bf2e3bad22c45dba24474377e.
  "Don't include sgame files from navgen code"
* d40e4fc7fdd19834094d367df9eb5e3fbeff88a1.
  "Shuffle definitions among some bot headers"
* 6849e625f967aecf482aecc46be32eb28cad676a.
  "Bots: turn MeshProcess into two classes"
* 8295d59e7a341860002fb27c8f1d3784e26b77dd.
  "Split navmesh generation into small chunks of work"
* fd8577d21942511d452f61824530f50fa7346ddf.
  "Implement 'navgen' command proof of concept"
* b2b61e76ad425ef5895cd912453a90959c8ae418.
  "Get navgen code to compile"
* 96fc3291988171f7049b104a06f3eda408c2d626.
  "Import unmodified navgen code from daemonmap"
* bot code requires MIN_WALK_NORMAL
