+++ CMakeLists.txt
@@ -95,7 +95,7 @@ if (BUILD_CGAME AND (BUILD_GAME_NATIVE_DLL OR BUILD_GAME_NATIVE_EXE OR NACL))
     include_directories(${RMLUI_INCLUDE_DIRS})
 endif()
 
-if (BUILD_SGAME OR BUILD_CGAME)
+if (BUILD_SGAME)
     # Fastlz
     add_library(srclibs-fastlz EXCLUDE_FROM_ALL ${FASTLZLIST})
     set_target_properties(srclibs-fastlz PROPERTIES POSITION_INDEPENDENT_CODE 1 FOLDER "libs")
@@ -107,13 +107,6 @@ if (BUILD_SGAME OR BUILD_CGAME)
                         ${LIB_DIR}/recastnavigation/DetourTileCache/Include
                         ${LIB_DIR}/recastnavigation/DebugUtils/Include
                         ${LIB_DIR}/recastnavigation/DetourCrowd/Include)
-
-    # Recast
-    add_library(srclibs-recast EXCLUDE_FROM_ALL ${RECASTLIST})
-    set_target_properties(srclibs-recast PROPERTIES POSITION_INDEPENDENT_CODE 1 FOLDER "libs")
-    include_directories(
-        ${LIB_DIR}/recastnavigation/Recast/Include
-        ${LIB_DIR}/recastnavigation/RecastDemo/Include)
 
     # GLM
     include_directories(${LIB_DIR}/glm)
@@ -165,7 +158,6 @@ if (BUILD_SGAME)
             ${sgame_GENERATED_CBSE}
         LIBS
             srclibs-detour
-            srclibs-recast
             srclibs-fastlz
     )
 endif()
@@ -184,8 +176,5 @@ if (BUILD_CGAME)
         ${FREETYPE_LIBRARIES}
         ${LUA_LIBRARY}
         srclibs-tinygettext
-        srclibs-detour
-        srclibs-recast
-        srclibs-fastlz
   )
 endif()
