+++ src/sgame/botlib/bot_local.h
@@ -45,7 +45,7 @@ ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 
 #include "bot_types.h"
 #include "bot_api.h"
-#include "shared/bot_nav_shared.h"
+#include "nav.h"
 #include "bot_convert.h"
 
 const int MAX_NAV_DATA = 16;
@@ -65,119 +65,13 @@ struct dtRouteResult
 	bool      invalid;
 };
 
-static const int NAVMESHCON_VERSION = 2;
-struct OffMeshConnectionHeader
-{
-	int version;
-	int numConnections;
-};
-
-struct OffMeshConnection
-{
-	rVec  start;
-	rVec  end;
-	float radius;
-	unsigned short flag;
-	unsigned char area;
-	unsigned char dir;
-	unsigned int userid;
-};
-
-struct OffMeshConnections
-{
-	static CONSTEXPR int MAX_CON = 128;
-	float  verts[ MAX_CON * 6 ]; // dtOffMeshConnection::pos
-	float  rad[ MAX_CON ]; // dtOffMeshConnection::rad
-	unsigned short flags[ MAX_CON ]; // dtOffMeshConnection::poly
-	unsigned char areas[ MAX_CON ]; // dtOffMeshConnection::flags
-	unsigned char dirs[ MAX_CON ]; // likely dtOffMeshConnection::side
-	unsigned int userids[ MAX_CON ]; // dtOffMeshConnection::userId
-	int      offMeshConCount;
-
-	OffMeshConnections() : offMeshConCount( 0 ) { }
-
-	void reset()
-	{
-		offMeshConCount = 0;
-	}
-
-	void delConnection( int index )
-	{
-		if ( index < 0 || index >= offMeshConCount )
-		{
-			return;
-		}
-
-		for ( int i = index * 6; i < offMeshConCount * 6 - 1; i++ )
-		{
-			verts[ i ] = verts[ i + 1 ];
-		}
-
-		for ( int i = index; i < offMeshConCount - 1; i++ )
-		{
-			rad[ i ] = rad[ i + 1 ];
-			flags[ i ] = flags[ i + 1 ];
-			areas[ i ] = areas[ i + 1 ];
-			dirs[ i ] = dirs[ i + 1 ];
-			userids[ i ] = userids[ i + 1 ];
-		}
-		offMeshConCount--;
-	}
-
-	bool addConnection( const OffMeshConnection &c )
-	{
-		if ( offMeshConCount == MAX_CON )
-		{
-			return false;
-		}
-
-		float *start = &verts[ offMeshConCount * 6 ];
-		float *end = start + 3;
-
-		start[ 0 ] = c.start[ 0 ];
-		start[ 1 ] = c.start[ 1 ];
-		start[ 2 ] = c.start[ 2 ];
-		end[ 0 ] = c.end[ 0 ];
-		end[ 1 ] = c.end[ 1 ];
-		end[ 2 ] = c.end[ 2 ];
-
-		rad[ offMeshConCount ] = c.radius;
-		flags[ offMeshConCount ] = c.flag;
-		areas[ offMeshConCount ] = c.area;
-		dirs[ offMeshConCount ] = c.dir;
-		userids[ offMeshConCount ] = c.userid;
-
-		offMeshConCount++;
-		return true;
-	}
-};
-
-struct NavconMeshProcess : public BasicMeshProcess
-{
-	OffMeshConnections con;
-
-	void process( struct dtNavMeshCreateParams *params, unsigned char *polyAreas, unsigned short *polyFlags ) override
-	{
-		// Update poly flags from areas.
-		BasicMeshProcess::process( params, polyAreas, polyFlags );
-
-		params->offMeshConVerts = con.verts;
-		params->offMeshConRad = con.rad;
-		params->offMeshConCount = con.offMeshConCount;
-		params->offMeshConAreas = con.areas;
-		params->offMeshConDir = con.dirs;
-		params->offMeshConFlags = con.flags;
-		params->offMeshConUserID = con.userids;
-	}
-};
-
 struct NavData_t
 {
 	dtTileCache      *cache;
 	dtNavMesh        *mesh;
 	dtNavMeshQuery   *query;
 	dtQueryFilter    filter;
-	NavconMeshProcess process;
+	MeshProcess      process;
 	char             name[ 64 ];
 };
