+++ src/sgame/sg_admin.h
@@ -168,7 +168,6 @@ bool        G_admin_l1( gentity_t *ent );  // AA-QVM 1.2
 bool        G_admin_register( gentity_t *ent );  // AA-QVM 1.2
 bool        G_admin_unregister( gentity_t *ent );  // AA-QVM 1.2
 bool        G_admin_bot( gentity_t *ent );
-bool        G_admin_navgen( gentity_t *ent );
 
 void            G_admin_print( gentity_t *ent, Str::StringRef m );
 void            G_admin_print_plural( gentity_t *ent, Str::StringRef m, int number );
