+++ src/sgame/sg_bot.cpp
@@ -546,21 +546,13 @@ void G_BotIntermissionThink( gclient_t *client )
 	client->readyToExit = true;
 }
 
-bool G_BotInit()
+void G_BotInit()
 {
-	if ( !G_BotNavInit() )
-	{
-		Log::Notice( "Failed to load navmeshes" );
-		G_BotNavCleanup();
-		return false;
-	}
-
+	G_BotNavInit( );
 	if ( treeList.maxTrees == 0 )
 	{
 		InitTreeList( &treeList );
 	}
-
-	return true;
 }
 
 void G_BotCleanup()
