+++ src/sgame/sg_bot_util.h
@@ -110,7 +110,7 @@ enum botMoveDir_t
 // global navigation
 extern bool navMeshLoaded;
 
-bool         G_BotNavInit();
+void         G_BotNavInit();
 void         G_BotNavCleanup();
 bool     FindRouteToTarget( gentity_t *self, botTarget_t target, bool allowPartial );
 bool         BotMoveToGoal( gentity_t *self );
